feat(Roster): Suggestion without context api
