--- packages/fluentui/docs/src/prototypes/Roster/actionsContext.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-export interface IActionsContext {
-  togglePromote: (id: string, type: string) => void;
-  toggleMute: (id: string, type: string) => void;
-}
-
-export const ActionsContext = React.createContext<IActionsContext>(undefined);
-
-export const useActions = () => React.useContext(ActionsContext);
